Max words, Mem is now reported correctly; unsure about Heap.
git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6640 c9681fa6-b8ae-40c0-a6da-707a35ad5973

+++ include/Sys/macros.h
@@ -89,7 +89,7 @@
 
 #define the_gc( globals )      ((gc_t*)globals[ G_GC ])
 
-#define bytes2words(x)      ((x) / sizeof(word))
+#define bytes2words(x)      (((unsigned) (x)) / sizeof(word))
 #define words2bytes(x)      ((x) * sizeof(word))
 
 /* Macros for manipulating Scheme data types. */
